Cleaned README and notebooks, restructured and renamed package
