+++ .gitignore
@@ -1,4 +1,4 @@
-.devcontainer/
+.devcontainer
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
@@ -162,7 +162,7 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
-
+*coverage*
 .cursorrules
 *.parquet
 *.jsonl
