+++ pyproject.toml
@@ -1,23 +1,58 @@
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
-
 [project]
-name = "time_series_trends"
-version = "0.1.0"
+name = "tstrends"
+dynamic = ["version", "dependencies", "classifiers"]
+description = "Advanced trend labelling for time series"
 authors = [
-    { name = "Your Name", email = "your.email@example.com" },
+    { name = "Abel G Penas", email = "agpenas.dev@gmail.com" }
 ]
-description = "A package for trend labelling in financial time series"
+maintainers = [
+    { name = "Abel G Penas", email = "agpenas.dev@gmail.com" }
+]
 readme = "README.md"
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: BSD-2-Clause License",
-    "Operating System :: OS Independent",
+packages = [
+    { include = "tstrends" }
+]
+license = {text = "BSD-2-Clause"}
+requires-python = ">=3.10"
+keywords = [
+    "time series",
+    "trend detection",
+    "forecasting",
+    "finantial tools",
+    "machine learning"
 ]
 
-[tool.setuptools.packages.find]
-where = ["time_series_trends"]
-include = ["trend_labelling*", "returns_estimation*", "optimization*"]
-namespaces = false 
+[project.urls]
+repository = "https://github.com/agpenas/tstrends"
+
+[tool.poetry]
+version = "0.1.6"
+classifiers = [
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Intended Audience :: Science/Research",
+    "Intended Audience :: Developers",
+    "Intended Audience :: Financial and Insurance Industry",
+]
+package-mode = true
+
+[tool.poetry.dependencies]
+python = "^3.10"
+matplotlib = "^3.10.0"
+bayesian-optimization = "^2.0.3"
+pandas = "^2.2.3"
+numpy = "^2.2.3"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.3.4"
+black = "^25.1.0"
+mypy = "^1.15.0"
+flake8 = "^7.1.2"
+pytest-cov = "^6.0.0"
+pytest-mock = "^3.14.0"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
