--- requirements-dev.txt
@@ -1,6 +1,0 @@
-pytest>=7.4.0
-pytest-cov>=4.1.0
-pytest-mock>=3.11.1
-black>=23.7.0
-mypy>=1.5.1
-flake8>=6.1.0
