--- requirements.txt
@@ -1,3 +1,0 @@
-numpy
-matplotlib
-bayesian-optimization>=2.0.0
