+++ tests/unit_tests/tstrends/trend_labelling/test_base_labeller.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 
-from time_series_trends.trend_labelling.base_labeller import BaseLabeller
-from time_series_trends.trend_labelling.label_scaling import Labels
+from tstrends.trend_labelling.base_labeller import BaseLabeller
+from tstrends.trend_labelling.label_scaling import Labels
 
 
 class ConcreteLabeller(BaseLabeller):
