+++ tstrends/trend_labelling/__init__.py
