+++ tstrends/trend_labelling/base_labeller.py
