+++ tstrends/trend_labelling/label_scaling.py
