+++ tstrends/trend_labelling/oracle_labeller.py
