+++ tstrends/trend_labelling/ternary_CTL.py
