+++ tstrends/visualization/__init__.py
@@ -1,0 +1,3 @@
+from .utils import plot_trend_labels
+
+__all__ = ["plot_trend_labels"]
